70. Resource based authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesForRestaurantCommandHandler.cs
@@ -4,11 +4,12 @@ using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishes
 {
-    public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> logger ,IRestaurantsRepository restaurantsRepository , IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesForRestaurantCommand>
+    public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> logger ,IRestaurantAuthorizationService restaurantAuthorizationService, IRestaurantsRepository restaurantsRepository , IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesForRestaurantCommand>
     {
         public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -18,7 +19,11 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishes
 
             if (restaurant == null) throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
 
-           await  dishesRepository.Delete(restaurant.Dishes);
+
+            if (!restaurantAuthorizationService.Authorize(restaurant, Domain.Constants.ResourceOperation.Update))
+                throw new ForbidException();
+
+            await  dishesRepository.Delete(restaurant.Dishes);
 
 
         }
